For Aesthetic purposes and I also changed the code
On the question that ask the uses for the license that they use, I notice that I made so the user can use multiple licenses for that, but reading on forums I realized that is not the way to do it since a rpogrammer can only do one license at a time and doing multimple licenses they would breach there "contract" with the license and also they are not compatiple with each other.  So instead I made it as a list type.  I also change it so the code it would not show the whole image file and instead show the name of the license.  As for the readme-template, I created a const generateLicense in which it would create the image so I could be transfered once you answer what license you want to choose.